Filter awareness in per-root-entity-vocab endpoint
