+++ build.gradle.kts
@@ -4,7 +4,7 @@ import java.net.URL
 
 plugins {
   java
-  id("org.veupathdb.lib.gradle.container.container-utils") version "4.8.9"
+  id("org.veupathdb.lib.gradle.container.container-utils") version "4.8.5"
   id("com.github.johnrengelman.shadow") version "7.1.2"
 }
 
@@ -87,8 +87,8 @@ repositories {
 
 // versions
 val coreLib       = "6.15.3"        // Container core lib version
-val edaCommon     = "10.7.1"        // EDA Common version
-val libSubsetting = "4.7.1"         // lib-eda-subsetting version
+val edaCommon     = "10.8.0"        // EDA Common version
+val libSubsetting = "4.8.0"         // lib-eda-subsetting version
 val fgputil       = "2.12.9-jakarta" // FgpUtil version
 
 // use local EdaCommon compiled schema if project exists, else use released version;
