+++ schema/library.raml
@@ -240,6 +240,11 @@ types:
       filters: API_Filter[]
       binSpec?: BinSpecWithRange
       valueSpec: ValueSpec
+  VocabByRootEntityPostRequest:
+    type: object
+    additionalProperties: false
+    properties:
+      filters: API_Filter[]
   VocabByRootEntityPostResponse:
     type: EntityTabularPostResponse
     additionalProperties: false
